Copy opencode.json from redsail-agents if an opencode.json file does not exist in current directory

+++ setup.sh
@@ -5,6 +5,11 @@
 cp -Rf redsail-agents/.agents ./.agents
 
 # Opencode
+if [ ! -f opencode.json ]; then
+  cp redsail-agents/opencode.json ./opencode.json
+  exit 0
+fi
+
 jq -s '.[0] * .[1]' redsail-agents/opencode.json opencode.json > opencode.json.tmp
 mv opencode.json.tmp opencode.json
